OpenNI2: Fix for slightly adjusted VideoPlaybackInterface

+++ src/video/drivers/openni2.cpp
@@ -26,11 +26,11 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-#include <pangolin/video/drivers/openni2.h>
 #include <pangolin/factory/factory_registry.h>
+#include <pangolin/video/drivers/openni2.h>
 
-#include <PS1080.h>
 #include <OniVersion.h>
+#include <PS1080.h>
 
 namespace pangolin
 {
@@ -209,7 +209,7 @@ void OpenNi2Video::InitialiseOpenNI()
     numDevices = 0;
     numStreams = 0;
     current_frame_index = 0;
-    total_frames = std::numeric_limits<int>::max();
+    total_frames = std::numeric_limits<size_t>::max();
 
     openni::Status rc = openni::STATUS_OK;
 
@@ -242,7 +242,7 @@ void OpenNi2Video::AddStream(const OpenNiStreamMode& mode)
 
     openni::PlaybackControl* control = device.getPlaybackControl();
     if(control && numStreams==0) {
-        total_frames = std::min(total_frames,control->getNumberOfFrames(stream));
+        total_frames = std::min(total_frames, (size_t)control->getNumberOfFrames(stream));
     }
 
     numStreams++;
@@ -610,17 +610,17 @@ bool OpenNi2Video::GrabNewest( unsigned char* image, bool wait )
     return GrabNext(image,wait);
 }
 
-int OpenNi2Video::GetCurrentFrameId() const
+size_t OpenNi2Video::GetCurrentFrameId() const
 {
     return current_frame_index;
 }
 
-int OpenNi2Video::GetTotalFrames() const
+size_t OpenNi2Video::GetTotalFrames() const
 {
     return total_frames;
 }
 
-int OpenNi2Video::Seek(int frameid)
+size_t OpenNi2Video::Seek(size_t frameid)
 {
     openni::PlaybackControl* control = devices[0].getPlaybackControl();
     if(control) {
